test(config): fix missing error path coverage for GPU detection
- Consolidate tests/test_config_coverage.py, tests/test_config_gguf.py,
  and tests/test_config_helpers.py into tests/test_config.py to adhere
  to the 1:1 mapping rule between src/ and tests/.
- Add comprehensive error path tests for `_detect_gpu` (ImportError
  simulated via sys.modules and Runtime Exception).
- Achieve 98% test coverage for mnemo_mcp.config.
- Clean up redundant test files.
