+++ tests/test_config.py
@@ -1,9 +1,16 @@
 """Tests for mnemo_mcp.config — Settings, API keys, embedding resolution."""
 
 import os
+import sys
 from pathlib import Path
+from unittest.mock import MagicMock, patch
 
-from mnemo_mcp.config import Settings
+from mnemo_mcp.config import (
+    Settings,
+    _detect_gpu,
+    _has_gguf_support,
+    _resolve_local_model,
+)
 
 
 class TestSettingsDefaults:
@@ -282,3 +289,281 @@ class TestGoogleDriveClientId:
         )
         s = Settings(api_keys=None)
         assert s.google_drive_client_id == "123456.apps.googleusercontent.com"
+
+
+# --- Consolidated from tests/test_config_coverage.py ---
+
+
+class TestSetupProviders:
+    def test_sdk_mode(self, monkeypatch):
+        """setup_providers configures SDK mode with API keys."""
+        s = Settings(
+            api_keys="GOOGLE_API_KEY:test-key",
+        )
+        mode = s.setup_providers()
+        assert mode == "sdk"
+        monkeypatch.delenv("GOOGLE_API_KEY", raising=False)
+        monkeypatch.delenv("GEMINI_API_KEY", raising=False)
+
+    def test_local_mode(self, monkeypatch):
+        """setup_providers returns 'local' when no API keys."""
+        for k in (
+            "JINA_AI_API_KEY",
+            "GEMINI_API_KEY",
+            "GOOGLE_API_KEY",
+            "OPENAI_API_KEY",
+            "COHERE_API_KEY",
+            "CO_API_KEY",
+            "XAI_API_KEY",
+        ):
+            monkeypatch.delenv(k, raising=False)
+        s = Settings(api_keys=None)
+        mode = s.setup_providers()
+        assert mode == "local"
+
+
+class TestResolveProviderMode:
+    def test_sdk_mode_with_api_keys(self):
+        """Returns 'sdk' when api_keys is set."""
+        s = Settings(api_keys="KEY:value")
+        assert s.resolve_provider_mode() == "sdk"
+
+    def test_local_mode_default(self, monkeypatch):
+        """Returns 'local' when nothing is configured."""
+        for k in (
+            "JINA_AI_API_KEY",
+            "GEMINI_API_KEY",
+            "GOOGLE_API_KEY",
+            "OPENAI_API_KEY",
+            "COHERE_API_KEY",
+            "CO_API_KEY",
+            "XAI_API_KEY",
+        ):
+            monkeypatch.delenv(k, raising=False)
+        s = Settings(api_keys=None)
+        assert s.resolve_provider_mode() == "local"
+
+    def test_sdk_mode_with_env_var(self, monkeypatch):
+        """Returns 'sdk' when a provider env var is set."""
+        monkeypatch.setenv("JINA_AI_API_KEY", "test-key")
+        s = Settings(api_keys=None)
+        assert s.resolve_provider_mode() == "sdk"
+
+    def test_sdk_mode_with_cohere_env(self, monkeypatch):
+        """Returns 'sdk' when COHERE_API_KEY env var is set."""
+        monkeypatch.setenv("COHERE_API_KEY", "test-key")
+        s = Settings(api_keys=None)
+        assert s.resolve_provider_mode() == "sdk"
+
+
+class TestResolveLocalEmbeddingModel:
+    def test_returns_onnx_by_default(self):
+        """Returns ONNX model when no GPU or no GGUF support."""
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=False),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=False),
+        ):
+            s = Settings(api_keys=None)
+            model = s.resolve_local_embedding_model()
+            assert "ONNX" in model
+
+    def test_returns_gguf_with_gpu_and_llama(self):
+        """Returns GGUF model when GPU is available and llama-cpp is installed."""
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=True),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            s = Settings(api_keys=None)
+            model = s.resolve_local_embedding_model()
+            assert "GGUF" in model
+
+
+class TestResolveEmbeddingBackendAuto:
+    def test_auto_detect_with_api_keys(self):
+        """Auto-detect returns 'cloud' when API keys are set."""
+        s = Settings(
+            api_keys="COHERE_API_KEY:test",
+            embedding_backend="",
+        )
+        assert s.resolve_embedding_backend() == "cloud"
+
+    def test_auto_detect_with_no_config(self, monkeypatch):
+        """Auto-detect returns 'local' when nothing is configured."""
+        for k in (
+            "JINA_AI_API_KEY",
+            "GEMINI_API_KEY",
+            "GOOGLE_API_KEY",
+            "OPENAI_API_KEY",
+            "COHERE_API_KEY",
+            "CO_API_KEY",
+            "XAI_API_KEY",
+        ):
+            monkeypatch.delenv(k, raising=False)
+        s = Settings(embedding_backend="", api_keys=None)
+        assert s.resolve_embedding_backend() == "local"
+
+    def test_litellm_backward_compat(self):
+        """'litellm' embedding_backend maps to 'cloud'."""
+        s = Settings(embedding_backend="litellm", api_keys=None)
+        assert s.resolve_embedding_backend() == "cloud"
+
+    def test_litellm_rerank_backward_compat(self):
+        """'litellm' rerank_backend maps to 'cloud'."""
+        s = Settings(rerank_backend="litellm", api_keys=None)
+        assert s.resolve_rerank_backend() == "cloud"
+
+
+class TestGpuDetectionErrorPaths:
+    def test_detect_gpu_import_error(self):
+        """_detect_gpu returns False when onnxruntime is not installed."""
+        with patch.dict(sys.modules, {"onnxruntime": None}):
+            assert _detect_gpu() is False
+
+    def test_detect_gpu_runtime_exception(self):
+        """_detect_gpu returns False when get_available_providers raises."""
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.side_effect = Exception("Runtime error")
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is False
+
+
+# --- Consolidated from tests/test_config_gguf.py ---
+
+
+def test_gguf_support_import_error():
+    import builtins
+
+    orig_import = builtins.__import__
+
+    def mock_import(name, *args, **kwargs):
+        if name == "llama_cpp":
+            raise ImportError("mocked")
+        return orig_import(name, *args, **kwargs)
+
+    with patch("builtins.__import__", side_effect=mock_import):
+        # We manually save and restore only the 'llama_cpp' entry
+        # rather than using patch.dict(sys.modules) which can cause
+        # side effects with other C extensions like numpy being
+        # unexpectedly unloaded when the dict is restored.
+        original_module = sys.modules.get("llama_cpp", None)
+        if "llama_cpp" in sys.modules:
+            del sys.modules["llama_cpp"]
+
+        try:
+            assert _has_gguf_support() is False
+        finally:
+            if original_module is not None:
+                sys.modules["llama_cpp"] = original_module
+
+
+# --- Consolidated from tests/test_config_helpers.py ---
+
+
+class TestDetectGPU:
+    def test_cuda_available(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = [
+            "CUDAExecutionProvider",
+            "CPUExecutionProvider",
+        ]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is True
+
+    def test_dml_available(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = [
+            "DmlExecutionProvider",
+            "CPUExecutionProvider",
+        ]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is True
+
+    def test_no_gpu_provider(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.return_value = ["CPUExecutionProvider"]
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is False
+
+    def test_import_error(self):
+        with patch.dict(sys.modules, {"onnxruntime": None}):
+            assert _detect_gpu() is False
+
+    def test_runtime_exception(self):
+        mock_ort = MagicMock()
+        mock_ort.get_available_providers.side_effect = Exception("Runtime error")
+        with patch.dict(sys.modules, {"onnxruntime": mock_ort}):
+            assert _detect_gpu() is False
+
+    def test_import_exception(self):
+        import builtins
+
+        orig_import = builtins.__import__
+
+        def mock_import(name, *args, **kwargs):
+            if name == "onnxruntime":
+                raise Exception("Mocked Exception")
+            return orig_import(name, *args, **kwargs)
+
+        with patch("builtins.__import__", side_effect=mock_import):
+            with patch.dict(sys.modules):
+                if "onnxruntime" in sys.modules:
+                    del sys.modules["onnxruntime"]
+                assert _detect_gpu() is False
+
+
+class TestHasGGUFSupport:
+    def test_llama_cpp_installed(self):
+        mock_llama = MagicMock()
+        with patch.dict(sys.modules, {"llama_cpp": mock_llama}):
+            assert _has_gguf_support() is True
+
+    def test_llama_cpp_missing(self):
+        with patch.dict(sys.modules, {"llama_cpp": None}):
+            assert _has_gguf_support() is False
+
+    def test_llama_cpp_import_error(self):
+        import builtins
+
+        orig_import = builtins.__import__
+
+        def mock_import(name, *args, **kwargs):
+            if name == "llama_cpp":
+                raise ImportError("Mocked ImportError")
+            return orig_import(name, *args, **kwargs)
+
+        with patch("builtins.__import__", side_effect=mock_import):
+            # Ensure sys.modules does not prevent the import call
+            original_module = sys.modules.get("llama_cpp", None)
+
+            if "llama_cpp" in sys.modules:
+                del sys.modules["llama_cpp"]
+
+            try:
+                assert _has_gguf_support() is False
+
+            finally:
+                if original_module is not None:
+                    sys.modules["llama_cpp"] = original_module
+
+
+class TestResolveLocalModel:
+    def test_gpu_and_gguf(self):
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=True),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            assert _resolve_local_model("onnx-model", "gguf-model") == "gguf-model"
+
+    def test_gpu_no_gguf(self):
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=True),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=False),
+        ):
+            assert _resolve_local_model("onnx-model", "gguf-model") == "onnx-model"
+
+    def test_no_gpu(self):
+        with (
+            patch("mnemo_mcp.config._detect_gpu", return_value=False),
+            patch("mnemo_mcp.config._has_gguf_support", return_value=True),
+        ):
+            assert _resolve_local_model("onnx-model", "gguf-model") == "onnx-model"
